backend/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
ffmpeg-python==0.2.*
flask-cors==4.*
Werkzeug==2.*
flask==2.*